requirements.txt
langchain==0.1.4
SQLAlchemy==2.0.25
pyodbc==5.0.1
pandas==2.2.0
plotly==5.18.0
sqlparse==0.4.4
azure-core==1.29.1
azure-data-tables==12.4.3
azure-storage-blob==12.19.0
python-dotenv==1.0.0
PyYAML==6.0
redis==5.0.1
streamlit==1.31.0
streamlit-authenticator==0.3.1
openai==1.10.0
tiktoken==0.5.2
langchain-community==0.0.17
langchain-openai==0.0.5
databricks-connect==14.3.0